+++ apps/taalwiz-api/src/user-preferences/dto/update-user-preferences.dto.ts
@@ -3,5 +3,5 @@ import { IsMongoId, IsOptional } from 'class-validator';
 export class UpdateUserPreferencesDto {
   @IsMongoId()
   @IsOptional()
-  currentBookmarkListId?: string;
+  currentVocabularyListId?: string;
 }