+++ apps/taalwiz-api/src/user-preferences/models/user-preferences.model.ts
@@ -2,7 +2,7 @@ import { InferSchemaType, model, Schema, Types } from 'mongoose';
 
 const UserPreferencesSchema = new Schema({
   userId: { type: Schema.Types.ObjectId, required: true, unique: true, ref: 'User' },
-  currentBookmarkListId: { type: Schema.Types.ObjectId, ref: 'BookmarkList', default: null },
+  currentVocabularyListId: { type: Schema.Types.ObjectId, ref: 'VocabularyList', default: null },
 });
 
 export type UserPreferencesDoc = InferSchemaType<typeof UserPreferencesSchema> & {