+++ apps/taalwiz-api/src/vocabulary/dto/create-vocabulary-list.dto.ts
@@ -1,6 +1,6 @@
 import { IsString, MinLength } from 'class-validator';
 
-export class CreateBookmarkListDto {
+export class CreateVocabularyListDto {
   @IsString()
   @MinLength(1)
   name = '';