+++ apps/taalwiz-api/src/vocabulary/dto/rename-vocabulary-list.dto.ts
@@ -1,6 +1,6 @@
 import { IsString, MinLength } from 'class-validator';
 
-export class RenameBookmarkListDto {
+export class RenameVocabularyListDto {
   @IsString()
   @MinLength(1)
   name = '';